align the naming of vqa_inference
